refactor(hash_join): Move JoinHashMap to separate mod (#15419)
* refactor(hash_join): Move JoinHashMap to separate mod

* Add description to join_hash_map module


---------
